Better allocate colors of lines.
Rather than using a d3.colorScale, we now have a pool of colors and assign
a color to each line's data when the data is prepared.

This means we can get the next color, release now unused colors (when a line
is deleted), and have the color data available for use when drawing the
line's key in mappiness.ui.