Update line data when edit form is submitted.
A bit trickier than it seemed like it would be at first.

All fine until we've passed the form data back to the controller and
then... we have to get the new values from the dataManager, but keep
the ID and line color from the existing line. Hadn't thought of that.

So now we set both of these as hidden elements in the form, and pass
that data back to the controller.

These can then be passed to the dataManager which, if they're present
will use them when creating the new returned data. If they're not
present, it'll generate a new ID and use the next color from the pool,
as normal.